action.php: Fix ACL permission check for namespace levels
In case the permission level from the nodisp tag is above page level,
we need to check the ACL permissions of the user in the current
namespace.

Otherwise for example in a case where the user has namespace
permissions and we use "<nodisp 4>", the user will never see
the content of the block as auth_quickaclcheck($page_id) will always
return a value of 2 or below.